# openapi.yaml
openapi: 3.0.1
info:
  title: WASL Admin Panle APIs
  description: "For this API, you need `Jwt` token in `Authorization` header, for authorization purposes. Example: `Authorization: Bearer <token>`"
  contact:
    email: omer.akram2012@gmail.com
  version: 1.0.0
...
servers:
  - url: "http://127.0.0.1:30001/"
    description: localhost server
  - url: "http://178.18.251.52:30001/"
    description: localhost server
...
paths:
...
  /auth/token:
    post:
      tags:
        - Token
      summary: "To obtain a {access_token, refresh_token} pair for accessing the API"
      operationId: /auth/token
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessToken"
        "400":
          description: Invalid username or password
          content: {}
        "405":
          description: Method not allowed
          content: {}
        "415":
          description: Unsupported Media Type
          content: {}
  /auth/token/refresh:
    post:
      tags:
        - Token
      summary: To refresh an expired (or not) access_token
      operationId: /auth/token/refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessToken"
        "400":
          description: Invalid refresh_token
          content: {}
        "405":
          description: Method not allowed
          content: {}
        "415":
          description: Unsupported Media Type
          content: {}
  /auth/forgotpwd:
    post:
      tags:
        - Token
      summary: To request a forgotten password
      operationId: /auth/forgotpwd
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordSuccessResponse"
        "400":
          description: Could not generate forgot password token.
          content: {}
        "405":
          description: Method not allowed
          content: {}
        "415":
          description: Unsupported Media Type
          content: {}
  /auth/forgotpwd/reset:
    post:
      tags:
        - Token
      summary: To reset a forgotten password
      operationId: /auth/forgotpwd/reset
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetForgotPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetForgotPasswordSuccessResponse"
        "400":
          description: Could not change password.
          content: {}
        "405":
          description: Method not allowed
          content: {}
        "415":
          description: Unsupported Media Type
          content: {}
...
  /categories/{categoryId}/sub-categories:
    get:
      tags:
        - Categories
      summary: Fetch sub categories of a category.
      operationId: getallsubcategories
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: string
          description: Parent category of current sub-category.
      responses:
        200:
          description: Ok.
      security:
        - bearerAuth: []  
    post:
      tags:
        - Categories
      summary: Add sub category.
      operationId: subcategory
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: string
          description: Parent category of current sub-category.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubCategorySchema"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubCategorySchema"
        "400":
          description: Bad Request.
          content: {}
        "403":
          description: Forbidden
          content: {}
        "405":
          description: Method not allowed
          content: {}
        "415":
          description: Unsupported Media Type
          content: {}
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    SubCategorySchema:
     type : object
     properties:
        name:
          type: string
...
    LoginRequest:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AccessToken:
      required:
        - access_token
        - refresh_token
        - ttl
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        ttl:
          type: integer
          format: int64
          description: Time in seconds for access_token expiry.
    TokenRefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
    ForgotPasswordSuccessResponse:
      type: object
      properties:
        message:
          type: string
...
    ResetForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
        forgot_pwd_token:
          type: string
          minLength: 36
          maxLength: 36
        new_password:
          type: string
        confirm_new_password:
          type: string
...
    ResetForgotPasswordSuccessResponse:
      type: object
      properties:
        message:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT